refactor(agent): remove visual artifact store
