+++ src/agent/internal/agent/function_agent.go
@@ -15,7 +15,6 @@ type FunctionAgent struct {
 	Tools             []langtools.Tool
 	OutputKey         string
 	ScreenshotPruning ScreenshotPruningConfig
-	VisualArtifacts   visualArtifactReader
 }
 
 type visualObservationTool interface {
@@ -328,10 +327,10 @@ func (a *FunctionAgent) visualScreenshotObservation(step schema.AgentStep) (visu
 	if !a.isVisualObservationTool(step.Action.Tool) {
 		return visualScreenshotObservation{}, false
 	}
-	return parseScreenshotObservation(step.Observation, a.VisualArtifacts)
+	return parseScreenshotObservation(step.Observation)
 }
 
-func parseScreenshotObservation(observation string, readers ...visualArtifactReader) (visualScreenshotObservation, bool) {
+func parseScreenshotObservation(observation string) (visualScreenshotObservation, bool) {
 	var result postActionScreenshotResult
 	if err := json.Unmarshal([]byte(observation), &result); err != nil {
 		return visualScreenshotObservation{}, false
@@ -339,20 +338,11 @@ func parseScreenshotObservation(observation string, readers ...visualArtifactRea
 	if result.Width <= 0 || result.Height <= 0 {
 		return visualScreenshotObservation{}, false
 	}
-	var imageBytes []byte
-	if strings.TrimSpace(result.Data) != "" {
-		var err error
-		imageBytes, err = base64.StdEncoding.DecodeString(result.Data)
-		if err != nil || len(imageBytes) == 0 {
-			return visualScreenshotObservation{}, false
-		}
-	} else if strings.TrimSpace(result.ScreenshotRef) != "" && len(readers) > 0 && readers[0] != nil {
-		var err error
-		imageBytes, err = readers[0].ReadVisualArtifact(result.ScreenshotRef)
-		if err != nil || len(imageBytes) == 0 {
-			return visualScreenshotObservation{}, false
-		}
-	} else {
+	if strings.TrimSpace(result.Data) == "" {
+		return visualScreenshotObservation{}, false
+	}
+	imageBytes, err := base64.StdEncoding.DecodeString(result.Data)
+	if err != nil || len(imageBytes) == 0 {
 		return visualScreenshotObservation{}, false
 	}
 	format, ok := normalizeScreenshotFormat(result.Format)
