+++ src/agent/internal/agent/task_episode.go
@@ -180,18 +180,6 @@ func (r *EpisodeRecorder) ID() string {
 	return r.id
 }
 
-func (r *EpisodeRecorder) visualArtifactRoot() string {
-	if r == nil {
-		return ""
-	}
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	if r.store == nil || strings.TrimSpace(r.store.rootDir) == "" {
-		return ""
-	}
-	return r.store.episodeDir(r.baseEpisodeLocked("running", time.Time{}))
-}
-
 func (r *EpisodeRecorder) Start(ctx context.Context) error {
 	if r == nil {
 		return nil
@@ -229,7 +217,6 @@ func (r *EpisodeRecorder) RecordDefaultFinish(answer string) {
 		Content: strings.TrimSpace(answer),
 	})
 }
-
 
 func (r *EpisodeRecorder) RecordPlannerExecution(result roleExecutionResult) {
 	r.recordExecutionForRole(result, RoleAgent)
