+++ src/agent/internal/agent/task_episode_error_test.go
@@ -1,6 +1,8 @@
 package agent
 
 import (
+	"path/filepath"
+	"strings"
 	"testing"
 
 	"github.com/tmc/langchaingo/schema"
@@ -35,4 +37,23 @@ func TestEpisodeRecorderRecordsStructuredToolError(t *testing.T) {
 		return
 	}
 	t.Fatal("missing tool_result event")
+}
+
+func TestMaterializeEventArtifactUsesExistingReference(t *testing.T) {
+	store := NewTaskEpisodeStore(t.TempDir())
+	dir := filepath.Join(t.TempDir(), "episode")
+	ref := "artifacts/step_002.jpeg"
+	event := TaskEpisodeEvent{
+		RawObservation: `{"width":320,"height":240,"format":"jpeg","size":10,"screenshot_ref":"` + ref + `"}`,
+	}
+
+	if err := store.materializeEventArtifact(dir, &event, 1); err != nil {
+		t.Fatalf("materialize referenced artifact: %v", err)
+	}
+	if event.ScreenshotRef != ref {
+		t.Fatalf("screenshot_ref = %q, want %q", event.ScreenshotRef, ref)
+	}
+	if !strings.Contains(event.Observation, `"screenshot_ref":"`+ref+`"`) {
+		t.Fatalf("compact observation is missing reference: %s", event.Observation)
+	}
 }
